Use global normalization for Tiff per propagation
Previously when storing tiffs a global nomalization factor
was applied to all frames whatever was the plane.
Now this factor depends on the plane. So there are as many
normalization factors as there are planes where the field
is propagated to.

+++ src/propagation/core/include/signal.hpp
@@ -5,12 +5,12 @@
 class Signal{
   public:
     Signal( double lambda, double side_length_in_meter, int N);
-    void detect( Eigen::MatrixXcd detecting, std::string filename, int max_value, int bit_depth);
-    void phase_detect( Eigen::MatrixXcd detecting, std::string filename, int max_value, int bit_depth);
-    void picture( std::string filename, int max_value, int bit_depth = 64);
+    void detect( Eigen::MatrixXcd detecting, std::string filename, double max_value, int bit_depth);
+    void phase_detect( Eigen::MatrixXcd detecting, std::string filename, double max_value, int bit_depth);
+    void picture( std::string filename, double max_value, int bit_depth = 64);
     double bucket();
     void bucket( std::string filename, double norm_fact);
-    void phase_picture( std::string filename, int max_value, int bit_depth);
+    void phase_picture( std::string filename, double max_value, int bit_depth);
     void toString();
     void quadratic_phase_lag_shift( double k, double c);
     void quadratic_phase_lag_shift_fourier( double k, double c);
@@ -19,11 +19,12 @@ class Signal{
     void propagate( double dist);
     void illuminate_thermally( double coherence_diameter);
     void illuminate_thermally_flat_knl( double coherence_diameter);
+    //void illuminate_thermally_fft( double coherence_diameter);
     void triple_slit_mask( int w_ratio, int h_ratio, int slits);
     double max_intensity();
   private:
     void illuminate_uniformly();
-    void store( std::string filename, Eigen::MatrixXd data_to_store, int max_value, int bit_depth);
+    void store( std::string filename, Eigen::MatrixXd data_to_store, double max_value, int bit_depth);
     double lambda;
     double L;
     Eigen::MatrixXcd value;
